pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "taller-mecanico-agent"
version = "0.1.0"
description = "Agent para atención al cliente en taller mecánico y venta de autopartes"
authors = [{name = "Tu Nombre", email = "tu@email.com"}]
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.0.60",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "psycopg2-binary>=2.9.0",
    "langchain-postgres>=0.0.1",
    "pydantic>=2.0.0",
    "langgraph-cli[inmem]>=0.4.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

